+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 Changelog
 =========
+## [0.4.0] - 2019-03-29
+### Added
+- Completion provider for model files
+- Reference contributor for model files
+- Plugin logo
+
 ## [0.3.0] - 2019-01-08
 ### Added
 - Completion provider for helper files
@@ -23,6 +29,7 @@ Changelog
 - Reference contributor for translations
 - Folding builder for translations
 
+[0.4.0]: https://github.com/martynassateika/CodeIgniter-phpstorm-plugin/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/martynassateika/CodeIgniter-phpstorm-plugin/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/martynassateika/CodeIgniter-phpstorm-plugin/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/martynassateika/CodeIgniter-phpstorm-plugin/compare/0.1.0...0.2.0